--- SharePointFramework/@Shared/src/services/ProjectDataService/IProjectDataServiceConfiguration.ts
@@ -1,9 +1,0 @@
-import { SpEntityPortalService } from 'sp-entityportal-service'
-
-export interface IProjectDataServiceConfiguration {
-  webUrl: string
-  spfxContext: any
-  siteId: string
-  entityService: SpEntityPortalService
-  propertiesListName: string
-}
