--- SharePointFramework/@Shared/src/services/ProjectDataService/IPropertyItemContext.ts
@@ -1,10 +1,0 @@
-import { IItem } from '@pnp/sp/items'
-import { IList } from '@pnp/sp/lists'
-
-export interface IPropertyItemContext {
-  itemId?: number
-  listId?: string
-  defaultEditFormUrl?: string
-  list?: IList
-  item?: IItem
-}
