+++ SharePointFramework/PortfolioWebParts/package.json
@@ -42,12 +42,12 @@
     "moment": "2.24.0",
     "msgraph-helper": "0.7.3",
     "object-assign": "4.1.1",
-    "pp365-shared": "1.7.3-892.2",
+    "pp365-shared": "1.7.3-891.5",
     "react": "16.13.1",
     "react-calendar-timeline": "0.27.0",
     "react-dom": "16.13.1",
     "react-highcharts": "16.0.2",
-    "sp-hubsite-service": "0.10.0",
+    "sp-hubsite-service": "0.10.1",
     "tslib": "2.3.1",
     "underscore": "1.9.1"
   },
