+++ SharePointFramework/ProjectWebParts/package.json
@@ -43,7 +43,7 @@
     "get-value": "3.0.1",
     "interactjs": "1.6.2",
     "msgraph-helper": "0.7.1",
-    "pp365-shared": "1.7.3-891.4",
+    "pp365-shared": "1.7.3-891.5",
     "pzl-spfx-components": "0.0.11",
     "react": "16.13.1",
     "react-calendar-timeline": "0.26.0",
@@ -53,7 +53,7 @@
     "react-scroll": "1.7.11",
     "rehype-raw": "^6.1.1",
     "sp-entityportal-service": "1.3.3",
-    "sp-hubsite-service": "0.10.0",
+    "sp-hubsite-service": "0.10.1",
     "spfx-jsom": "^0.6.6",
     "tslib": "2.3.1",
     "underscore": "1.9.1"
